+++ app/src/main/java/com/example/housekeeper/costomView/model/Item.kt
@@ -3,6 +3,6 @@ package com.example.housekeeper.costomView.model
 import com.example.housekeeper.domain.model.Expense
 
 data class Item(
-    val value: Long,
+    val value: Double,
     val category: Expense
 )
